Use Upsource API to get/show review title
Use the Upsource API to get the review title and show it in the Slack
message board. The review title provides very useful context to Slack
users, which is otherwise missing. Unfortunately the title is not
available in the event bean, so a new request has to be made to obtain
it.

Many optimizations and improvements are possible:

* caching to avoid repeated requests for the same info
* use of review details in other ways e.g. differentiate the types of
participants